refactor: get image through node instead of fetching against the server
